+++ src/torch_uncertainty/models/segmentation/unet/packed.py
@@ -16,7 +16,7 @@ class _PackedDoubleConv(nn.Module):
         self,
         in_channels: int,
         out_channels: int,
-        alpha: int,
+        alpha: float,
         num_estimators: int,
         gamma: int,
         mid_channels: int | None = None,
@@ -58,7 +58,7 @@ class _PackedDoubleConv(nn.Module):
 
 class _PackedInconv(nn.Module):
     def __init__(
-        self, in_channels: int, out_channels: int, alpha: int, num_estimators: int, gamma: int
+        self, in_channels: int, out_channels: int, alpha: float, num_estimators: int, gamma: int
     ) -> None:
         super().__init__()
         self.conv = _PackedDoubleConv(
@@ -76,7 +76,7 @@ class _PackedInconv(nn.Module):
 
 class _PackedDown(nn.Module):
     def __init__(
-        self, in_channels: int, out_channels: int, alpha: int, num_estimators: int, gamma: int
+        self, in_channels: int, out_channels: int, alpha: float, num_estimators: int, gamma: int
     ) -> None:
         super().__init__()
         self.mpconv = nn.Sequential(
@@ -101,7 +101,7 @@ class _PackedUp(nn.Module):
         self,
         in_channels: int,
         out_channels: int,
-        alpha: int,
+        alpha: float,
         num_estimators: int,
         gamma: int,
         bilinear: bool = True,
@@ -151,7 +151,7 @@ class _PackedUp(nn.Module):
 
 class _PackedOutconv(nn.Module):
     def __init__(
-        self, in_channels: int, out_channels: int, alpha: int, num_estimators: int, gamma: int
+        self, in_channels: int, out_channels: int, alpha: float, num_estimators: int, gamma: int
     ) -> None:
         super().__init__()
         self.conv = PackedConv2d(
@@ -174,7 +174,7 @@ class _PackedUNet(nn.Module):
         in_channels: int,
         num_classes: int,
         num_blocks: list[int],
-        alpha: int = 1,
+        alpha: float = 1.0,
         num_estimators: int = 1,
         gamma: int = 1,
         bilinear: bool = False,
@@ -243,7 +243,7 @@ def _packed_unet(
     num_classes: int,
     num_blocks: list[int],
     bilinear: bool = False,
-    alpha: int = 1,
+    alpha: float = 1.0,
     num_estimators: int = 1,
     gamma: int = 1,
     dropout_rate: float = 0.0,
@@ -279,7 +279,7 @@ def packed_small_unet(
     in_channels: int,
     num_classes: int,
     bilinear: bool = False,
-    alpha: int = 1,
+    alpha: float = 1.0,
     num_estimators: int = 1,
     gamma: int = 1,
     dropout_rate: float = 0.0,
@@ -316,7 +316,7 @@ def packed_unet(
     in_channels: int,
     num_classes: int,
     bilinear: bool = False,
-    alpha: int = 1,
+    alpha: float = 1.0,
     num_estimators: int = 1,
     gamma: int = 1,
     dropout_rate: float = 0.0,
